+++ integration/src/test/java/de/digitalcollections/flusswerk/integration/DeadletteringTest.java
@@ -1,4 +1,5 @@
-package de.digitalcollections.flusswerk.bdd;
+package de.digitalcollections.flusswerk.integration;
+
 
 import de.digitalcollections.flusswerk.engine.Engine;
 import de.digitalcollections.flusswerk.engine.flow.Flow;
@@ -14,7 +15,17 @@ import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class SuccessFulProcessingTest {
+/**
+ * Feature: Dead letter messages
+ *
+ *   Scenario: Message processing fails constantly
+ *     Given I have an message broker with default config and a message in integration.in
+ *     When the processing always fails
+ *     Then the message in queue integration.in.failed has 5 retries
+ *      And integration.in is empty
+ *      And integration.out is empty
+ */
+public class DeadletteringTest {
 
   private final static String QUEUE_IN = "test.in";
   private final static String QUEUE_OUT = "test.out";
@@ -29,12 +40,14 @@ public class SuccessFulProcessingTest {
   }
 
   @Test
-  public void successfulMessagesShouldGoToOutQueue() throws Exception {
+  public void failingMessagesShouldEndInFailedQueue() throws Exception {
     MessageBroker messageBroker = backend.getMessageBroker();
-    Flow flow = new FlowBuilder<DefaultMessage, DefaultMessage, DefaultMessage>()
-        .read(m -> m)
-        .transform(m -> m)
-        .writeAndSend((DefaultMessage m) -> m)
+    Flow flow = new FlowBuilder<>()
+        .read(Message::toString)
+        .transform(s -> {
+          throw new RuntimeException("Fail!");
+        })
+        .write(s -> new DefaultMessage(s.toString()))
         .build();
 
     Engine engine = new Engine(messageBroker, flow);
@@ -42,11 +55,12 @@ public class SuccessFulProcessingTest {
     ExecutorService executorService = Executors.newSingleThreadExecutor();
     executorService.submit(engine::start);
     messageBroker.send(QUEUE_IN, new DefaultMessage("123456"));
+//    Thread.sleep(5000);
 
-    Message message = backend.waitForMessageFrom(QUEUE_OUT, 10_000);
+    Message message = backend.waitForMessageFrom(QUEUE_FAILED, 10_000);
     assertThat(message.getId()).isEqualTo("123456");
     assertThatMessageFrom(QUEUE_IN).isNull();
-    assertThatMessageFrom(QUEUE_FAILED).isNull();
+    assertThatMessageFrom(QUEUE_OUT).isNull();
     assertThatMessageFrom(QUEUE_RETRY).isNull();
 
     engine.stop();
@@ -58,5 +72,3 @@ public class SuccessFulProcessingTest {
   }
 
 }
-
-
